Revert "kernel: Fix all GCC 7.x warnings"
This reverts commit 5a16c51f17df810b7885a22acf27ef268cdb66f3.

+++ drivers/usb/gadget/function/f_fs.c
@@ -1716,7 +1716,7 @@ static void functionfs_unbind(struct ffs_data *ffs)
 static int ffs_epfiles_create(struct ffs_data *ffs)
 {
 	struct ffs_epfile *epfile, *epfiles;
-	short i, count;
+	unsigned i, count;
 
 	ENTER();
